+++ octobot-packages/OctoBot-Tentacles-Manager/octobot_tentacles_manager/api/__init__.py
@@ -63,6 +63,7 @@ from octobot_tentacles_manager.api.configurator import (
     set_tentacles_setup_configuration_path,
     ensure_setup_configuration,
     refresh_profile_tentacles_setup_config,
+    refresh_all_tentacles_setup_configs,
     get_code_hash,
     get_config_hash,
 )
@@ -72,6 +73,7 @@ from octobot_tentacles_manager.api.updater import (
 )
 from octobot_tentacles_manager.api.loader import (
     load_tentacles,
+    is_tentacles_architecture_valid,
     reload_tentacle_info,
     ensure_tentacle_info,
     register_extra_tentacle_data,
@@ -93,6 +95,9 @@ from octobot_tentacles_manager.api.inspector import (
     check_tentacle_version,
     get_tentacle_class_from_string,
     get_tentacles_classes_names_for_type,
+    get_installed_packages_from_url,
+    get_tentacles_from_package_name,
+    get_all_installed_package_urls,
 )
 from octobot_tentacles_manager.api.uninstaller import (
     uninstall_all_tentacles,
@@ -137,9 +142,11 @@ __all__ = [
     "get_config_hash",
     "ensure_setup_configuration",
     "refresh_profile_tentacles_setup_config",
+    "refresh_all_tentacles_setup_configs",
     "update_all_tentacles",
     "update_tentacles",
     "load_tentacles",
+    "is_tentacles_architecture_valid",
     "reload_tentacle_info",
     "ensure_tentacle_info",
     "register_extra_tentacle_data",
@@ -158,6 +165,9 @@ __all__ = [
     "check_tentacle_version",
     "get_tentacle_class_from_string",
     "get_tentacles_classes_names_for_type",
+    "get_installed_packages_from_url",
+    "get_tentacles_from_package_name",
+    "get_all_installed_package_urls",
     "uninstall_all_tentacles",
     "uninstall_tentacles",
     "start_tentacle_creator",