+++ profiler/start_view_backtesting_profiler_result.sh
@@ -1,0 +1,3 @@
+# install snakeviz first
+# python -m pip install snakeviz
+snakeviz result-backtesting.prof