+++ octobot-packages/trading-backend/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.25] - 2024-07-05
+### Updated
+- [CCXT] update to ccxt 4.3.56
+
 ## [1.2.24] - 2024-06-16
 ### Updated
 - Exchanges: fix api key format errors